Fixed this file
Note: Testing for this isn't quite working. There appears to be a break between the integration of the splitter (which is basically a multiplexer), and the actual output when it's connected to the rest. The splitter tests appear fine. All the Roms are fine. Holder appears to be fine too. So it is likely an issue with addr